requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.34
python-multipart==0.0.9
pillow==10.4.0
pytesseract==0.3.13
